requirements.txt
altgraph==0.17
beautifulsoup4==4.9.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
chromedriver==2.24.1
click==7.1.2
comtypes==1.1.7
configparser==5.0.0
cryptography==2.9.2
cx-oracle-on-ctypes==0.1.dev0
dotnetAddDlls==0.1
ffmpeg==1.4
ffmpeg-python==0.2.0
future==0.18.2
gTTS==2.1.1
gTTS-token==1.1.3
idna==2.9
importlib-metadata==1.7.0
inflect==4.1.0
itsdangerous==1.1.0
jaraco.itertools==5.0.0
joblib==0.15.1
lxml==4.5.1
more-itertools==8.4.0
MouseInfo==0.1.3
nltk==3.5
numpy==1.19.0
opencv-python==4.3.0.36
panda==0.3.1
pdfminer.six==20200517
pefile==2019.4.18
Pillow==7.1.2
pip==20.1.1
playsound==1.2.2
py4j==0.10.9
PyAudio==0.2.11
PyAutoGUI==0.9.50
pycparser==2.20
pycryptodome==3.9.8
pydotnet==0.9.2
pydub==0.24.1
PyGetWindow==0.0.8
PyInstaller==3.6
PyMsgBox==1.0.8
pyperclip==1.8.0
PyRect==0.1.4
PyScreeze==0.1.26
pytesseract==0.3.4
pythonnet==2.5.1
pytils==0.3
pyttsx3==2.88
PyTweening==1.0.3
pywin32==228
pywin32-ctypes==0.2.0
pywinauto==0.6.8
PyYAML==5.3.1
regex==2020.6.8
requests==2.24.0
scikit-learn==0.23.1
scipy==1.5.0
seamless==0.1.0
selenium==3.141.0
setuptools==44.0.0
six==1.15.0
sklearn==0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SpeechRecognition==3.8.1
threadpoolctl==2.1.0
tools==0.1.9
tqdm==4.46.1
urllib3==1.25.9
Wand==0.6.2
wolframalpha==4.0.0
xmltodict==0.12.0
zipp==3.1.0
